delete operation added on mylist section

+++ src/pages/MyList/MyList.jsx
@@ -2,6 +2,7 @@ import { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../../providers/AuthProvider';
 import { MdDeleteForever } from 'react-icons/md';
 import { MdSystemUpdateAlt } from 'react-icons/md';
+import Swal from 'sweetalert2';
 
 const MyList = () => {
   const { user } = useContext(AuthContext) || {};
@@ -16,9 +17,33 @@ const MyList = () => {
       });
   }, [user]);
 
-  const handleDelete = email => {
-    fetch(`http://localhost:5000/myList/${email}`, {
-      method: 'DELETE',
+  const handleDelete = id => {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then(result => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/allTouristsSpot/${id}`, {
+          method: 'DELETE',
+        })
+          .then(res => res.json())
+          .then(data => {
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: 'Deleted!',
+                text: 'Your file has been deleted.',
+                icon: 'success',
+              });
+              const remaining = myTouristsSpot.filter(item => item._id !== id);
+              setMyTouristsSpot(remaining);
+            }
+          });
+      }
     });
   };
 
@@ -62,7 +87,7 @@ const MyList = () => {
                 <td className="px-6 py-4">{item?.cost}</td>
                 <td className="px-6 py-4">
                   <button
-                    onClick={() => handleDelete(user?.email)}
+                    onClick={() => handleDelete(item._id)}
                     className="btn btn-error mr-5"
                   >
                     Delete
